Apps/requirements.txt
psycopg2
redis
python-dotenv